newscrawler/requirements.txt
celery==3.1.23
cffi==1.8.2
cryptography==1.5
cssselect==0.9.2
Django==1.10.1
html5lib==1.0b2
Jinja2==2.8
lxml==3.6.4
parsel==1.0.3
Pillow==3.3.1
pycparser==2.14
PyPDF2==1.26.0
python-dateutil==2.5.3
redis==2.10.5
w3lib==1.15.0
webencodings==0.5
xhtml2pdf==0.0.6
